+++ src/header/manager.ts
@@ -44,9 +44,12 @@ export class HeaderManager {
     this.searchBar.remove()
   }
 
-  public closeButtonClick (): void {
+  public closeButtonClick (to: any): void {
     this.showSearch()
-    this._map.footerManager.showVenue().catch((): void => null)
+    if (to && to.objectClass === 'place') {
+      return this._map.footerManager.setSelected(to)
+    }
+    return this._map.footerManager.showVenue().catch((): void => null)
   }
 
   public showSearch (): Promise<void> {
@@ -91,10 +94,10 @@ export class HeaderManager {
 
   public search (searchString: string, searchOptions: any, clickOnResultCallback: (searchResult: any, universe?: any) => void, focusedField: string): void {
     this.searchResults.showLoading()
-    const transformedSearchQuery = callOptionnalFn(this._options.onSearchQueryWillBeSent, [searchString, searchOptions])
-    search(transformedSearchQuery.searchString, transformedSearchQuery.searchOptions).then((searchResults: any) => {
+    const transformedSearchOptions = callOptionnalFn(this._options.onSearchQueryWillBeSent, [searchOptions, searchString, focusedField])
+    search(searchString, transformedSearchOptions).then((searchResults: any) => {
       this.searchResults.hideLoading()
-      this.showSearchResults(callOptionnalFn(this._options.onSearchResultWillBeDisplayed, [searchResults]), clickOnResultCallback, focusedField)
+      this.showSearchResults(callOptionnalFn(this._options.onSearchResultsWillBeDisplayed, [searchResults]), clickOnResultCallback, focusedField)
     })
   }