+++ src/header/searchBar/searchBar.ts
@@ -22,12 +22,19 @@ export class SearchBar extends DefaultControl {
 
   private _hideSearchResultsTimeout: any
   private _currentVenueState: number
+  private _hideMenu: boolean
 
   constructor (mapInstance: any, options: any) {
     super(mapInstance)
 
     this._options = options
     this._container = $('<div />')
+
+    if (!isFunction(this._options.onMenuButtonClick)) {
+      this._hideMenu = true
+    } else {
+      this._hideMenu = false
+    }
 
     this.listen('click', '#mwz-menu-button', this._menuButtonClick.bind(this))
     this.listen('click', '#mwz-header-directions-button', this._directionButtonClick.bind(this))
@@ -56,17 +63,17 @@ export class SearchBar extends DefaultControl {
 
   public enteringIn (venue: any): void {
     this._currentVenueState = ENTERING_IN_VENUE
-    this._container.html(template(enteringInVenueHtml)({ hide_menu: this._options.hideMenu, entering_in: translate('entering_in_venue', { venue: getTranslation(venue, this.map.getLanguageForVenue(venue._id), 'title') }) }))
+    this._container.html(template(enteringInVenueHtml)({ hide_menu: this._hideMenu, entering_in: translate('entering_in_venue', { venue: getTranslation(venue, this.map.getLanguageForVenue(venue._id), 'title') }) }))
   }
   public enteredIn (venue: any): void {
     this._currentVenueState = IN_VENUE
-    this._container.html(template(inVenueHtml)({ hide_menu: this._options.hideMenu, search_in: translate('search_placeholder_venue', { venue: getTranslation(venue, this.map.getLanguageForVenue(venue._id), 'title') }) }))
+    this._container.html(template(inVenueHtml)({ hide_menu: this._hideMenu, search_in: translate('search_placeholder_venue', { venue: getTranslation(venue, this.map.getLanguageForVenue(venue._id), 'title') }) }))
     this.refreshTooltips()
   }
   public leaveVenue (): void {
     this.refreshTooltips()
     this._currentVenueState = OUT_OF_VENUE
-    this._container.html(template(outOfVenueHtml)({ hide_menu: this._options.hideMenu, search_in: translate('search_placeholder_global') }))
+    this._container.html(template(outOfVenueHtml)({ hide_menu: this._hideMenu, search_in: translate('search_placeholder_global') }))
   }
 
   public refreshLocale () {
@@ -101,7 +108,7 @@ export class SearchBar extends DefaultControl {
       setTimeout(() => {
         this._container.find('#mwz-menu-button').attr('data-original-title', translate('menu'))
         this._container.find('#mwz-header-directions-button').attr('data-original-title', translate('directions'))
-        this._container.find('[data-toggle="tooltip"]').tooltip({ trigger: 'hover', container: this.map._container })
+        this._container.find('[data-toggle="mwz-tooltip"]').tooltip({ trigger: 'hover', container: this.map._container })
       }, 1000)
     }
   }
@@ -151,24 +158,15 @@ export class SearchBar extends DefaultControl {
     }, 500)
   }
 
-  private _clickOnSearchResult (searchResult: any, universe?: any): void {
-    if (this._map.getVenue()) {
-      this._map.footerManager.setSelected(searchResult)
-    }
-    this._centerOnSearchResult(searchResult, universe)
-  }
-
-  private _centerOnSearchResult (searchResult: any, universe?: any): void {
+  private _clickOnSearchResult (searchResult: any, universe?: any, analytics: any = null): void {
     if (searchResult._id) {
-      const venue = this._map.getVenue()
-      if (universe && venue) {
-        this.map.setUniverseForVenue(universe._id, venue)
-      }
-
       if (searchResult.objectClass === 'venue') {
-        return this.map.centerOnVenue(searchResult._id)
-      } else if (searchResult.objectClass === 'place') {
-        return this.map.centerOnPlace(searchResult._id)
+        this.map.centerOnVenue(searchResult._id)
+      } else if (searchResult.objectClass === 'place' || searchResult.objectClass === 'placeList') {
+        if (universe) {
+          this.map.setUniverseForVenue(universe._id, searchResult.venueId)
+        }
+        this._map.footerManager.setSelected(searchResult, true, analytics)
       }
     } else if (searchResult.geometry) {
       if (searchResult.geometry.bounds) {